requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.7.0
requests==2.31.0

# AI/LLM Integration
openai==1.35.0
anthropic==0.28.0

# Data Processing
pandas==2.2.0
numpy==1.26.0

# WebSocket for real-time data
websockets==12.0
aiohttp==3.9.0

# API Framework
fastapi==0.110.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0

